Improve fp8 experiments

+++ experiments/experiment.py
@@ -18,7 +18,7 @@ BATCH_SIZE = 10000
 DIM = 4096
 
 # Datatype for matmul + activation caching: torch.bfloat16 or torch.float8_e4m3fn.
-DTYPE = torch.float8_e4m3fn # torch.bfloat16 #
+DTYPE = torch.bfloat16 # torch.float8_e4m3fn #
 
 # Correctness tolerance: exact for bf16, loose for the fp8 quantization error.
 CHECK_RTOL = CHECK_ATOL = 3e-6 if DTYPE == torch.bfloat16 else 1e-1
